Finished styling for the Dashboard.
