+++ src/components/pages/Dashboard/RenderDashboard.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { nav } from '../../../config';
 import { Header, SubCard } from '../../common';
+import PromptBox from './PromptBox';
 
 interface RenderDashboardProps {
   picList: { src: string; alt?: string }[];
@@ -9,6 +10,12 @@ interface RenderDashboardProps {
 const RenderDashboard = ({
   picList,
 }: RenderDashboardProps): React.ReactElement => {
+  // The following values should be loaded from BE
+  const prompt = {
+    username: 'CatLady',
+    prompt: 'You are a super hero, and there is a town that needs saving.',
+    streak: 3,
+  };
   return (
     <div className="dashboard-container">
       <Header menuItems={nav.siteNavItems} />
@@ -30,37 +37,6 @@ const RenderDashboard = ({
       </div>
     </div>
   );
-};
-
-const PromptBox = (props: PromptBoxProps): React.ReactElement => {
-  return (
-    <div className="prompt-box">
-      <h2>Hey, {props.username}!</h2>
-      <h3>Here is today&apos;s prompt:</h3>
-      <p>{props.prompt}</p>
-      <div className="prompt-footer">
-        {/* <div className="streak">
-          <h3>Hot Streak:</h3>
-          <span className="flames">
-            {[...new Array(props.streak)].map(() => '(f)')}
-          </span>
-        </div> */}
-        <button>Submit Your Story</button>
-      </div>
-    </div>
-  );
-};
-
-interface PromptBoxProps {
-  username: string;
-  prompt: string;
-  streak: number;
-}
-
-const prompt = {
-  username: 'CatLady',
-  prompt: 'You are a super hero, and there is a town that needs saving.',
-  streak: 3,
 };
 
 export default RenderDashboard;
